Refactor PreparationFeature.swift into separate files
