+++ MomentumApp/Sources/PreparationFeature.swift
@@ -4,24 +4,6 @@ import Sharing
 
 @Reducer
 struct PreparationFeature {
-    struct ChecklistSlot: Equatable, Identifiable, Codable {
-        let id: Int // Position 0-3
-        var item: ChecklistItem?
-        var isTransitioning: Bool = false
-        var isFadingIn: Bool = false
-        
-        private enum CodingKeys: String, CodingKey {
-            case id, item
-            // Don't persist animation states
-        }
-    }
-    
-    struct ItemTransition: Equatable {
-        let slotId: Int
-        let replacementText: String?
-        let startTime: Date
-    }
-    
     @ObservableState
     struct State: Equatable {
         var goal: String = ""
@@ -128,92 +110,34 @@ struct PreparationFeature {
                 return .none
                 
             case let .checklistSlotToggled(slotId):
-                guard slotId < state.checklistSlots.count,
-                      let item = state.checklistSlots[slotId].item else { return .none }
-                
-                if !item.isCompleted {
-                    // Mark item as completed
-                    var slots = state.checklistSlots
-                    slots[slotId].item?.isCompleted = true
-                    state.checklistSlots = slots
-                    state.totalItemsCompleted += 1
-                    
-                    // Check if we have more items to show
-                    let replacementText: String?
-                    if state.nextItemIndex < ChecklistItemPool.allItems.count {
-                        replacementText = ChecklistItemPool.allItems[state.nextItemIndex]
-                        state.nextItemIndex += 1
-                    } else {
-                        replacementText = nil
-                    }
-                    
-                    // Start fade-out transition after 600ms delay
-                    return .run { send in
-                        try await clock.sleep(for: .milliseconds(600))
-                        await send(.beginSlotTransition(slotId: slotId, replacementText: replacementText))
-                    }
-                } else {
-                    // Unchecking an item (not in spec, but handling for completeness)
-                    var slots = state.checklistSlots
-                    slots[slotId].item?.isCompleted = false
-                    state.checklistSlots = slots
-                    state.totalItemsCompleted -= 1
-                }
-                return .none
-                
-            case let .beginSlotTransition(slotId, replacementText):
-                // Mark slot as transitioning
-                var slots = state.checklistSlots
-                slots[slotId].isTransitioning = true
-                state.checklistSlots = slots
-                state.activeTransitions[slotId] = ItemTransition(
+                return Self.handleChecklistSlotToggled(
+                    state: &state,
                     slotId: slotId,
-                    replacementText: replacementText,
-                    startTime: Date()
+                    clock: clock
                 )
                 
-                // Complete transition after fade-out duration (300ms)
-                return .run { send in
-                    try await clock.sleep(for: .milliseconds(300))
-                    await send(.completeSlotTransition(slotId: slotId))
-                }
+            case let .beginSlotTransition(slotId, replacementText):
+                return Self.handleBeginSlotTransition(
+                    state: &state,
+                    slotId: slotId,
+                    replacementText: replacementText,
+                    clock: clock
+                )
                 
             case let .completeSlotTransition(slotId):
-                guard let transition = state.activeTransitions[slotId] else { return .none }
-                
-                // First, clear the slot completely to prevent overlap
-                var slots = state.checklistSlots
-                slots[slotId].item = nil
-                slots[slotId].isTransitioning = false
-                slots[slotId].isFadingIn = false
-                state.checklistSlots = slots
-                
-                state.activeTransitions.removeValue(forKey: slotId)
-                
-                // If there's a replacement, add it after a small gap
-                if let replacementText = transition.replacementText {
-                    return .run { send in
-                        // Small delay to ensure clean visual gap
-                        try await clock.sleep(for: .milliseconds(100))
-                        await send(.fadeInNewItem(slotId: slotId, text: replacementText))
-                    }
-                }
-                return .none
+                return Self.handleCompleteSlotTransition(
+                    state: &state,
+                    slotId: slotId,
+                    clock: clock
+                )
                 
             case let .fadeInNewItem(slotId, text):
-                // Add the new item to the slot with fade-in animation
-                let newId = UUID().uuidString
-                let newItem = ChecklistItem(id: newId, text: text, isCompleted: false)
-                var slots = state.checklistSlots
-                slots[slotId].item = newItem
-                slots[slotId].isFadingIn = true
-                state.checklistSlots = slots
-                
-                // Reset the fade-in flag after animation duration
-                return .run { send in
-                    try await clock.sleep(for: .milliseconds(350)) // Slightly longer than animation
-                    await send(.resetFadeInFlag(slotId: slotId))
-                }
+                return Self.handleFadeInNewItem(
+                    state: &state,
+                    slotId: slotId,
+                    text: text,
+                    clock: clock
+                )
                 
             case let .resetFadeInFlag(slotId):
                 // Reset the fade-in flag so the item becomes fully interactive
